code/langgraph/requirements.txt
langchain-tavily>=0.2.13
langchain-openai>=1.0.2
langgraph>=1.0.3
python-dotenv>=1.2.1
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.12.4
